Project discontinuation
I've decided that it's best that I come back to this project several months to a year or two later.

Reason being that it is certainly do-able, but I just don't want the entire game's functionality to only work due to janky code. For this reason I've decided to put aside this project for now, and work on enhancing my current and learning new skills, until the day I choose to come back to this.